+++ CHANGELOG.md
@@ -5,6 +5,8 @@
 + Add `spec.parallism` field to limit concurrent pod execution at a workflow level
 + Add `template.parallism` field to limit concurrent pod execution at a template level
 + Add `argo pause`, `argo resume` to pause and resume workflows
++ Add `argo resubmit` to resubmit a failed workflow (EXPERIMENTAL)
++ Improved parameters and output validation
 
 ## 2.0.0 (2018-02-06)
 + Add ability to specify affinity rules at both the workflow and template level